Undo wrong approach for element_time_derivative
Previous approach was correct for mass_residual but not for
element_time_derivative. That's going to be more involved for the
Rayleigh damping.
