+++ src/physics/src/elastic_cable.C
@@ -30,6 +30,7 @@
 #include "grins/solid_mechanics_bc_handling.h"
 #include "grins/generic_ic_handler.h"
 #include "grins/postprocessed_quantities.h"
+#include "grins/elasticity_tensor.h"
 
 // libMesh
 #include "libmesh/getpot.h"
@@ -223,6 +224,149 @@ namespace GRINS
               }
           }
       }
+  }
+
+  template<typename StressStrainLaw>
+  void ElasticCable<StressStrainLaw>::element_time_derivative( bool compute_jacobian,
+                                                               AssemblyContext& context,
+                                                               CachedValues& /*cache*/)
+  {
+    const unsigned int n_u_dofs = context.get_dof_indices(this->_disp_vars.u()).size();
+
+    const std::vector<libMesh::Real> &JxW =
+      this->get_fe(context)->get_JxW();
+
+    // Residuals that we're populating
+    libMesh::DenseSubVector<libMesh::Number> &Fu = context.get_elem_residual(this->_disp_vars.u());
+    libMesh::DenseSubVector<libMesh::Number> &Fv = context.get_elem_residual(this->_disp_vars.v());
+    libMesh::DenseSubVector<libMesh::Number> &Fw = context.get_elem_residual(this->_disp_vars.w());
+
+    //Grab the Jacobian matrix as submatrices
+    //libMesh::DenseMatrix<libMesh::Number> &K = context.get_elem_jacobian();
+    libMesh::DenseSubMatrix<libMesh::Number> &Kuu = context.get_elem_jacobian(this->_disp_vars.u(),this->_disp_vars.u());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kuv = context.get_elem_jacobian(this->_disp_vars.u(),this->_disp_vars.v());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kuw = context.get_elem_jacobian(this->_disp_vars.u(),this->_disp_vars.w());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kvu = context.get_elem_jacobian(this->_disp_vars.v(),this->_disp_vars.u());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kvv = context.get_elem_jacobian(this->_disp_vars.v(),this->_disp_vars.v());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kvw = context.get_elem_jacobian(this->_disp_vars.v(),this->_disp_vars.w());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kwu = context.get_elem_jacobian(this->_disp_vars.w(),this->_disp_vars.u());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kwv = context.get_elem_jacobian(this->_disp_vars.w(),this->_disp_vars.v());
+    libMesh::DenseSubMatrix<libMesh::Number> &Kww = context.get_elem_jacobian(this->_disp_vars.w(),this->_disp_vars.w());
+
+
+    unsigned int n_qpoints = context.get_element_qrule().n_points();
+
+    // All shape function gradients are w.r.t. master element coordinates
+    const std::vector<std::vector<libMesh::Real> >& dphi_dxi = this->get_fe(context)->get_dphidxi();
+
+    const libMesh::DenseSubVector<libMesh::Number>& u_coeffs = context.get_elem_solution( this->_disp_vars.u() );
+    const libMesh::DenseSubVector<libMesh::Number>& v_coeffs = context.get_elem_solution( this->_disp_vars.v() );
+    const libMesh::DenseSubVector<libMesh::Number>& w_coeffs = context.get_elem_solution( this->_disp_vars.w() );
+
+    // Need these to build up the covariant and contravariant metric tensors
+    const std::vector<libMesh::RealGradient>& dxdxi  = this->get_fe(context)->get_dxyzdxi();
+
+    const unsigned int dim = 1; // The cable dimension is always 1 for this physics
+
+    for (unsigned int qp=0; qp != n_qpoints; qp++)
+      {
+        // Gradients are w.r.t. master element coordinates
+        libMesh::Gradient grad_u, grad_v, grad_w;
+
+        for( unsigned int d = 0; d < n_u_dofs; d++ )
+          {
+            libMesh::RealGradient u_gradphi( dphi_dxi[d][qp] );
+            grad_u += u_coeffs(d)*u_gradphi;
+            grad_v += v_coeffs(d)*u_gradphi;
+            grad_w += w_coeffs(d)*u_gradphi;
+          }
+
+        libMesh::RealGradient grad_x( dxdxi[qp](0) );
+        libMesh::RealGradient grad_y( dxdxi[qp](1) );
+        libMesh::RealGradient grad_z( dxdxi[qp](2) );
+
+        libMesh::TensorValue<libMesh::Real> a_cov, a_contra, A_cov, A_contra;
+        libMesh::Real lambda_sq = 0;
+
+        this->compute_metric_tensors( qp, *(this->get_fe(context)), context,
+                                      grad_u, grad_v, grad_w,
+                                      a_cov, a_contra, A_cov, A_contra,
+                                      lambda_sq );
+
+        // Compute stress tensor
+        libMesh::TensorValue<libMesh::Real> tau;
+        ElasticityTensor C;
+        this->_stress_strain_law.compute_stress_and_elasticity(dim,a_contra,a_cov,A_contra,A_cov,tau,C);
+
+
+        libMesh::Real jac = JxW[qp];
+
+        for (unsigned int i=0; i != n_u_dofs; i++)
+          {
+            libMesh::RealGradient u_gradphi( dphi_dxi[i][qp] );
+
+            const libMesh::Real res_term = tau(0,0)*this->_A*jac*u_gradphi(0);
+
+            Fu(i) += res_term*(grad_x(0) + grad_u(0));
+
+            Fv(i) += res_term*(grad_y(0) + grad_v(0));
+
+            Fw(i) += res_term*(grad_z(0) + grad_w(0));
+          }
+
+        if( compute_jacobian )
+          {
+            for(unsigned int i=0; i != n_u_dofs; i++)
+              {
+                libMesh::RealGradient u_gradphi_I( dphi_dxi[i][qp] );
+                for(unsigned int j=0; j != n_u_dofs; j++)
+                  {
+                    libMesh::RealGradient u_gradphi_J( dphi_dxi[j][qp] );
+
+                    const libMesh::Real diag_term = this->_A*jac*tau(0,0)*( u_gradphi_J(0)*u_gradphi_I(0))*context.get_elem_solution_derivative();
+
+                    Kuu(i,j) += diag_term;
+
+                    Kvv(i,j) += diag_term;
+
+                    Kww(i,j) += diag_term;
+
+                    const libMesh::Real dgamma_du = ( u_gradphi_J(0)*(grad_x(0)+grad_u(0)) );
+
+                    const libMesh::Real dgamma_dv = ( u_gradphi_J(0)*(grad_y(0)+grad_v(0)) );
+
+                    const libMesh::Real dgamma_dw = ( u_gradphi_J(0)*(grad_z(0)+grad_w(0)) );
+
+                    const libMesh::Real C1 = this->_A*jac*C(0,0,0,0)*context.get_elem_solution_derivative();
+
+                    const libMesh::Real x_term = C1*( (grad_x(0)+grad_u(0))*u_gradphi_I(0) );
+
+                    const libMesh::Real y_term = C1*( (grad_y(0)+grad_v(0))*u_gradphi_I(0) );
+
+                    const libMesh::Real z_term = C1*( (grad_z(0)+grad_w(0))*u_gradphi_I(0) );
+
+                    Kuu(i,j) += x_term*dgamma_du;
+
+                    Kuv(i,j) += x_term*dgamma_dv;
+
+                    Kuw(i,j) += x_term*dgamma_dw;
+
+                    Kvu(i,j) += y_term*dgamma_du;
+
+                    Kvv(i,j) += y_term*dgamma_dv;
+
+                    Kvw(i,j) += y_term*dgamma_dw;
+
+                    Kwu(i,j) += z_term*dgamma_du;
+
+                    Kwv(i,j) += z_term*dgamma_dv;
+
+                    Kww(i,j) += z_term*dgamma_dw;
+
+                  } // end j-loop
+              } // end i-loop
+          } // end if(compute_jacobian)
+      } // end qp loop
   }
 
 } // end namespace GRINS
